fix: apply stderr snapshot gap and split oversized tests
Render stderr written between the primary and secondary log-stream
snapshots so the live logs pane does not drop that window. Split
bootstrap and dashboard CLI test files under the 300-line gate.

Generated-By: looper 0.0.0-dev (runner=fixer, agent=grok-build)
