+++ internal/api/bootstrap_test.go
@@ -158,244 +158,6 @@ func TestBootstrapConcurrentExchangeOnlyOneWins(t *testing.T) {
 	}
 }
 
-func TestOriginMismatchRejectedOnUnsafeMethod(t *testing.T) {
-	t.Parallel()
-
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/loops", bytes.NewReader([]byte(`{}`)))
-	req.Host = "127.0.0.1:17310"
-	req.Header.Set("Origin", "http://evil.example")
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusForbidden {
-		t.Fatalf("status = %d, want 403", rec.Code)
-	}
-}
-
-func TestOriginMatchAllowedOnUnsafeMethod(t *testing.T) {
-	t.Parallel()
-
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	// Unknown route still proves auth/origin passed (not 403).
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/does-not-exist", bytes.NewReader([]byte(`{}`)))
-	req.Host = "127.0.0.1:17310"
-	req.Header.Set("Origin", "http://127.0.0.1:17310")
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code == http.StatusForbidden {
-		t.Fatalf("status = 403, want non-forbidden (got body %s)", rec.Body.String())
-	}
-}
-
-func TestCLIWithoutOriginAllowed(t *testing.T) {
-	t.Parallel()
-
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/does-not-exist", bytes.NewReader([]byte(`{}`)))
-	req.Host = "127.0.0.1:17310"
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code == http.StatusForbidden {
-		t.Fatalf("CLI without Origin must not get 403")
-	}
-}
-
-func TestAttackerHostAndOriginRejectedAuthNone(t *testing.T) {
-	t.Parallel()
-
-	// DNS rebinding: Host and Origin both match the attacker domain — must still 403.
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/does-not-exist", bytes.NewReader([]byte(`{}`)))
-	req.Host = "evil.example"
-	req.Header.Set("Origin", "http://evil.example")
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusForbidden {
-		t.Fatalf("status = %d, want 403 body=%s", rec.Code, rec.Body.String())
-	}
-}
-
-func TestAttackerHostAndOriginRejectedOnSafeGetAuthNone(t *testing.T) {
-	t.Parallel()
-
-	// DNS rebinding via browser-readable GET (config / logs / status): Host+Origin
-	// from the attacker domain must 403 even when authMode=none. Auth runs before
-	// route dispatch, so unknown paths still prove the guard.
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	for _, path := range []string{"/api/v1/config", "/api/v1/status", "/api/v1/does-not-exist"} {
-		req := httptest.NewRequest(http.MethodGet, path, nil)
-		req.Host = "evil.example"
-		req.Header.Set("Origin", "http://evil.example")
-		rec := httptest.NewRecorder()
-		h.ServeHTTP(rec, req)
-		if rec.Code != http.StatusForbidden {
-			t.Fatalf("GET %s status = %d, want 403 body=%s", path, rec.Code, rec.Body.String())
-		}
-	}
-}
-
-func TestAttackerHostWithoutOriginRejectedOnSafeGetAuthNone(t *testing.T) {
-	t.Parallel()
-
-	// Same-origin DNS rebinding omits Origin; Host allowlist must still reject.
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	for _, path := range []string{"/api/v1/config", "/api/v1/status", "/api/v1/does-not-exist"} {
-		req := httptest.NewRequest(http.MethodGet, path, nil)
-		req.Host = "evil.example"
-		rec := httptest.NewRecorder()
-		h.ServeHTTP(rec, req)
-		if rec.Code != http.StatusForbidden {
-			t.Fatalf("GET %s without Origin status = %d, want 403 body=%s", path, rec.Code, rec.Body.String())
-		}
-	}
-}
-
-func TestCLIGetWithoutOriginAllowed(t *testing.T) {
-	t.Parallel()
-
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	// Unknown route still proves auth/origin passed (not 403).
-	req := httptest.NewRequest(http.MethodGet, "/api/v1/does-not-exist", nil)
-	req.Host = "127.0.0.1:17310"
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code == http.StatusForbidden {
-		t.Fatalf("CLI GET without Origin must not get 403")
-	}
-}
-
-func TestLegitimateHostOriginAllowedOnSafeGet(t *testing.T) {
-	t.Parallel()
-
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	req := httptest.NewRequest(http.MethodGet, "/api/v1/does-not-exist", nil)
-	req.Host = "127.0.0.1:17310"
-	req.Header.Set("Origin", "http://127.0.0.1:17310")
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code == http.StatusForbidden {
-		t.Fatalf("legitimate loopback GET Host/Origin must not get 403: %s", rec.Body.String())
-	}
-}
-
-func TestLegitimateHostOriginAllowed(t *testing.T) {
-	t.Parallel()
-
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:     "127.0.0.1",
-			Port:     17310,
-			AuthMode: config.AuthModeNone,
-		},
-	}})
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/does-not-exist", bytes.NewReader([]byte(`{}`)))
-	req.Host = "localhost:17310"
-	req.Header.Set("Origin", "http://127.0.0.1:17310")
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code == http.StatusForbidden {
-		t.Fatalf("legitimate loopback Host/Origin must not get 403: %s", rec.Body.String())
-	}
-}
-
-func TestHTTPSBaseURLBareHostAllowed(t *testing.T) {
-	t.Parallel()
-
-	// Remote HTTPS baseURL: browser sends Host without :443 and Origin without port.
-	baseURL := "https://dashboard.example.com"
-	token := "secret-token"
-	h := NewHandler(Context{Config: config.Config{
-		Server: config.ServerConfig{
-			Host:       "0.0.0.0",
-			Port:       17310,
-			AuthMode:   config.AuthModeLocalToken,
-			LocalToken: &token,
-			BaseURL:    &baseURL,
-		},
-	}})
-
-	req := httptest.NewRequest(http.MethodPost, "/api/v1/does-not-exist", bytes.NewReader([]byte(`{}`)))
-	req.Host = "dashboard.example.com"
-	req.Header.Set("Origin", "https://dashboard.example.com")
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", "Bearer "+token)
-	rec := httptest.NewRecorder()
-	h.ServeHTTP(rec, req)
-
-	if rec.Code == http.StatusForbidden {
-		t.Fatalf("bare Host with https baseURL must not get 403: %s", rec.Body.String())
-	}
-}
-
 func TestBootstrapAuthFailureSetsNoStore(t *testing.T) {
 	t.Parallel()
 
@@ -433,69 +195,6 @@ func TestBootstrapAuthFailureSetsNoStore(t *testing.T) {
 	}
 	if got := rec.Header().Get("Cache-Control"); got != "no-store" {
 		t.Fatalf("exchange host reject Cache-Control = %q, want no-store", got)
-	}
-}
-
-func TestNewRootHandlerDashboardMethodNotAllowed(t *testing.T) {
-	t.Parallel()
-
-	root := NewRootHandler(http.NotFoundHandler(), http.NotFoundHandler())
-	rec := httptest.NewRecorder()
-	root.ServeHTTP(rec, httptest.NewRequest(http.MethodPost, "/dashboard", nil))
-	if rec.Code != http.StatusMethodNotAllowed {
-		t.Fatalf("status = %d, want 405", rec.Code)
-	}
-	if got := rec.Header().Get("Allow"); got != "GET, HEAD" {
-		t.Fatalf("Allow = %q, want GET, HEAD", got)
-	}
-}
-
-func TestNewRootHandlerDashboardRedirectAndAPI(t *testing.T) {
-	t.Parallel()
-
-	apiHits := 0
-	api := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		apiHits++
-		w.WriteHeader(http.StatusTeapot)
-	})
-	dash := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write([]byte("dash"))
-	})
-	root := NewRootHandler(api, dash)
-
-	rec := httptest.NewRecorder()
-	root.ServeHTTP(rec, httptest.NewRequest(http.MethodGet, "/dashboard", nil))
-	if rec.Code != http.StatusFound {
-		t.Fatalf("redirect status = %d, want 302", rec.Code)
-	}
-	if loc := rec.Header().Get("Location"); loc != "/dashboard/" {
-		t.Fatalf("Location = %q, want /dashboard/", loc)
-	}
-
-	rec = httptest.NewRecorder()
-	root.ServeHTTP(rec, httptest.NewRequest(http.MethodGet, "/dashboard/", nil))
-	if rec.Code != http.StatusOK || rec.Body.String() != "dash" {
-		t.Fatalf("dashboard serve = %d %q", rec.Code, rec.Body.String())
-	}
-
-	rec = httptest.NewRecorder()
-	root.ServeHTTP(rec, httptest.NewRequest(http.MethodGet, "/api/v1/status", nil))
-	if rec.Code != http.StatusTeapot || apiHits != 1 {
-		t.Fatalf("api status = %d hits=%d", rec.Code, apiHits)
-	}
-
-	// Host-root favicon is served from embedded dashboard assets (not the API).
-	rec = httptest.NewRecorder()
-	root.ServeHTTP(rec, httptest.NewRequest(http.MethodGet, "/favicon.ico", nil))
-	if rec.Code != http.StatusOK {
-		t.Fatalf("favicon status = %d, want 200", rec.Code)
-	}
-	if rec.Body.Len() == 0 {
-		t.Fatal("favicon body empty")
-	}
-	if apiHits != 1 {
-		t.Fatalf("favicon must not hit API handler, apiHits=%d", apiHits)
 	}
 }
